+++ app/src/main/java/com/cointrendnotifier/android/api/dtos/EventDto.java
@@ -3,11 +3,11 @@ package com.cointrendnotifier.android.api.dtos;
 public class EventDto {
     private String _id;
     private double probability;
-    private int firedAt;
+    private long firedAt;
     private String baseAssetName;
     private String quoteAssetName;
 
-    public EventDto(String baseAssetName, String quoteAssetName, double probability, int firedAt, String _id) {
+    public EventDto(String baseAssetName, String quoteAssetName, double probability, long firedAt, String _id) {
         this._id = _id;
         this.probability = probability;
         this.firedAt = firedAt;
@@ -23,7 +23,7 @@ public class EventDto {
         return probability;
     }
 
-    public int getFiredAt() {
+    public long getFiredAt() {
         return firedAt;
     }
